+++ tests/test_m4i.py
@@ -1,6 +1,5 @@
 import unittest
 
-
 from ontolutils.ex.m4i import TextVariable, NumericalVariable
 
 
@@ -16,12 +15,11 @@ class TestM4i(unittest.TestCase):
 
     def testNumericalVariableWithoutStandardName(self):
         numerical_variable = NumericalVariable(
-            hasUnit='Unit',
+            hasUnit='m/s',
             hasNumericalValue=1.0,
             hasMaximumValue=2.0,
             hasVariableDescription='Variable description')
-        self.assertEqual(numerical_variable.hasUnit, 'Unit')
+        self.assertEqual(numerical_variable.hasUnit, 'http://qudt.org/vocab/unit/M-PER-SEC')
         self.assertEqual(numerical_variable.hasNumericalValue, 1.0)
         self.assertEqual(numerical_variable.hasMaximumValue, 2.0)
         self.assertEqual(numerical_variable.hasVariableDescription, 'Variable description')
-