+++ tests/test_serialization.py
@@ -4,6 +4,9 @@ import unittest
 import rdflib
 
 from ontolutils import Thing
+from ontolutils import serialize
+from ontolutils.ex import hdf5
+from ontolutils.ex import m4i
 
 LOG_LEVEL = logging.DEBUG
 
@@ -58,3 +61,23 @@ class TestSerialization(unittest.TestCase):
         res = g.query(_QUERY)
         bindings = res.bindings[0]
         self.assertEqual(bindings[rdflib.Variable("s")], rdflib.URIRef("https://example.com/123"))
+
+    def test_serialize_mutliple_things(self):
+        thing1 = m4i.Tool(
+            id="http://example.org/tool1",
+            label="Tool"
+        )
+        thing2 = hdf5.Dataset(
+            id="http://example.org/tool1",
+            name="/ds"
+        )
+        self.assertEqual(serialize([thing1, thing2], "ttl"), """@prefix hdf5: <http://purl.allotrope.org/ontologies/hdf5/1.8#> .
+@prefix m4i: <http://w3id.org/nfdi4ing/metadata4ing#> .
+@prefix rdfs: <http://www.w3.org/2000/01/rdf-schema#> .
+
+<http://example.org/tool1> a hdf5:Dataset,
+        m4i:Tool ;
+    rdfs:label "Tool" ;
+    hdf5:name "/ds" .
+
+""")