+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - bugfixes
 - improved DCAT classes
+- add `serialization()` function
 
 ## v0.22.0-rc2